some renaming and store update
