--- src/components/pages/Triggers/TriggersList/AddTriggersModal/index.ts
@@ -1,1 +1,0 @@
-export {default} from './AddTriggerModal';
